requirements.txt
Django==2.1.3
djangorestframework==3.9.0
graphene-django>=2.0
django-filter
django-graphql-jwt